Refactor. Not use bitrix_gt24.sh anymore. Switch to ansible-roles

+++ install_full_environment_fpm.sh
@@ -8,12 +8,10 @@ set -euo pipefail
 # bash <(curl -sL https://raw.githubusercontent.com/YogSottot/DebianLikeBitrixVM/feature/php-fpm/install_full_environment_fpm.sh)
 
 # use wget
+# apt install wget -y
+# wget https://raw.githubusercontent.com/YogSottot/DebianLikeBitrixVM/feature/php-fpm/.env.menu.example -O /root/.env.menu
+# edit .env.menu with your settings.
 # bash <(wget -qO- https://raw.githubusercontent.com/YogSottot/DebianLikeBitrixVM/feature/php-fpm/install_full_environment_fpm.sh)
-
-cat > /root/temp_install_full_environment.sh <<\END
-#!/usr/bin/env bash
-set +x
-set -euo pipefail
 
 generate_password() {
     local length=$1
@@ -26,96 +24,28 @@ generate_password() {
         password+=$char
     done
 
-    echo $password
+    echo "$password"
 }
 
 BRANCH="feature/php-fpm"
-SETUP_BITRIX_DEBIAN_URL="https://raw.githubusercontent.com/YogSottot/DebianLikeBitrixVM/$BRANCH/repositories/bitrix-gt/bitrix24_gt.sh"
 REPO_URL="https://github.com/YogSottot/DebianLikeBitrixVM"
 
 DB_NAME="bitrix"
 DB_USER="bitrix"
+DBPASS=$(generate_password 24)
 
 DIR_NAME_MENU="vm_menu"
 DEST_DIR_MENU="/root"
 
 FULL_PATH_MENU_FILE="$DEST_DIR_MENU/$DIR_NAME_MENU/menu.sh"
 
-# Function to compare versions
-version_gt() { test "$(printf '%s\n' "$@" | sort -V | head -n 1)" != "$1"; }
-
-# Get OS and version information
-if [ -f /etc/os-release ]; then
-    . /etc/os-release
-    OS=$NAME
-    VER=$VERSION_ID
-elif [ -f /etc/lsb-release ]; then
-    . /etc/lsb-release
-    OS=$DISTRIB_ID
-    VER=$DISTRIB_RELEASE
-fi
-
-# Check if it's Ubuntu 22.04 or less
-if [[ "$OS" == "Ubuntu" && $(echo -e "$VER\n22.04" | sort -V | head -n1) == "$VER" ]]; then
-    apt update -y
-    apt install -y software-properties-common
-    add-apt-repository -y ppa:ondrej/nginx
-fi
 apt update -y
 apt upgrade -y
-apt install -y perl wget curl ansible git ssl-cert cron locales locales-all poppler-utils catdoc nginx-light libnginx-mod-http-brotli-filter libnginx-mod-http-brotli-static libnginx-mod-http-headers-more-filter unattended-upgrades software-properties-common webp optipng jpegoptim ghostscript gifsicle
+apt install -y ansible git
+# fix for mysql role
+ansible-galaxy collection install 'community.mysql:==3.10.3'
 
-# set timezone
-timedatectl set-timezone Europe/Moscow
-
-# Set locales
-locale-gen en_US.UTF-8
-locale-gen en_GB.UTF-8
-locale-gen en_DK.UTF-8
-locale-gen ru_RU.UTF-8
-locale-gen C.UTF-8
-
-cat > /etc/default/locale <<CONFIG_LOCALE
-LANG=en_GB.UTF-8
-LANGUAGE=
-LC_CTYPE="en_GB.UTF-8"
-LC_NUMERIC="ru_RU.UTF-8"
-LC_TIME="en_DK.UTF-8"
-LC_COLLATE="ru_RU.UTF-8"
-LC_MONETARY="C.UTF-8"
-LC_MESSAGES="en_GB.UTF-8"
-LC_PAPER="ru_RU.UTF-8"
-LC_NAME="ru_RU.UTF-8"
-LC_ADDRESS="ru_RU.UTF-8"
-LC_TELEPHONE="en_GB.UTF-8"
-LC_MEASUREMENT="C.UTF-8"
-LC_IDENTIFICATION="ru_RU.UTF-8"
-LC_ALL=
-CONFIG_LOCALE
-
-source /etc/default/locale
-export LC_ALL="en_US.UTF-8"
-
-bash -c "$(curl -sL $SETUP_BITRIX_DEBIAN_URL)"
-
-source /root/run.sh
-
-set +x
-set -euo pipefail
-
-# set mysql root password
-#root_pass=$(generate_password 24)
 site_user_password=$(generate_password 24)
-
-#mysql -e "SET PASSWORD FOR 'root'@'localhost' = PASSWORD('${root_pass}');FLUSH PRIVILEGES;"
-
-#cat > /root/.my.cnf <<CONFIG_MYSQL_ROOT_MY_CNF
-#[client]
-#user=root
-#password="${root_pass}"
-# socket=/var/lib/mysqld/mysqld.sock
-
-#CONFIG_MYSQL_ROOT_MY_CNF
 
 # Clone directory vm_menu with repositories
 git clone --branch=$BRANCH --depth 1 --filter=blob:none --sparse $REPO_URL "$DEST_DIR_MENU/DebianLikeBitrixVM"
@@ -123,9 +53,9 @@ cd "$DEST_DIR_MENU/DebianLikeBitrixVM"
 git sparse-checkout set $DIR_NAME_MENU
 
 # Move vm_menu in /root and clean
-rm -rf $DEST_DIR_MENU/$DIR_NAME_MENU
+rm -rf "${DEST_DIR_MENU:?}/${DIR_NAME_MENU:?}"
 mv -f $DIR_NAME_MENU $DEST_DIR_MENU
-rm -rf "$DEST_DIR_MENU/DebianLikeBitrixVM"
+rm -rf "${DEST_DIR_MENU:?}/DebianLikeBitrixVM"
 
 cd $DEST_DIR_MENU
 
@@ -142,73 +72,86 @@ if ! grep -qF "$FULL_PATH_MENU_FILE" /root/.profile; then
 INSTALL_MENU
 fi
 
-# Configure apache2 modules
-a2enmod remoteip
-a2enmod rewrite
-a2enmod setenvif
-a2dismod ssl
-
-cat > /etc/apache2/mods-enabled/remoteip.conf <<CONFIG_APACHE2_REMOTEIP
-<IfModule remoteip_module>
-  RemoteIPHeader X-Real-IP
-  RemoteIPInternalProxy 127.0.0.1
-</IfModule>
-CONFIG_APACHE2_REMOTEIP
-
-cat > /etc/apache2/conf-enabled/php.conf <<CONFIG_APACHE2_FASTCGI
-      #
-      # The following lines prevent .user.ini files from being viewed by Web clients.
-      #
-      <Files ".user.ini">
-          <IfModule mod_authz_core.c>
-              Require all denied
-          </IfModule>
-          <IfModule !mod_authz_core.c>
-              Order allow,deny
-              Deny from all
-              Satisfy All
-          </IfModule>
-      </Files>
-
-      # Cause the PHP interpreter to handle files with a .php extension.
-      <FilesMatch "\.php$">
-      #        SetHandler "proxy:fcgi://127.0.0.1:9000"
-              SetHandler "proxy:unix:/run/php/php-fpm.sock|fcgi://localhost"
-      #       AddType application/x-httpd-php .php
-      </FilesMatch>
-
-      # Add index.php to the list of files that will be served as directory
-      # indexes.
-
-      DirectoryIndex index.php
-
-      # Uncomment the following line to allow PHP to pretty-print .phps
-      # files as PHP source code:
-      #
-      #AddType application/x-httpd-php-source .phps
-CONFIG_APACHE2_FASTCGI
-
-# set PHP 8.2
-update-alternatives --set php /usr/bin/php8.2
-update-alternatives --set phar /usr/bin/phar8.2
-update-alternatives --set phar.phar /usr/bin/phar.phar8.2
-update-alternatives --set php-fpm.sock /run/php/php8.2-fpm.sock
-
-
-ln -s $FULL_PATH_MENU_FILE "$DEST_DIR_MENU/menu.sh"
+ln -fs $FULL_PATH_MENU_FILE "$DEST_DIR_MENU/menu.sh"
 
 # Final actions
-
+# shellcheck source=/dev/null
 source $DEST_DIR_MENU/$DIR_NAME_MENU/bash_scripts/config.sh
 
-DOCUMENT_ROOT="${BS_PATH_SITES}/bx-site"
+# shellcheck source=/dev/null
+if [ -e /root/.env.menu ]; then
+  source /root/.env.menu
+fi
 
-DELETE_FILES=(
-  "$BS_PATH_APACHE_SITES_CONF/000-default.conf"
-  "$BS_PATH_APACHE_SITES_ENABLED/000-default.conf"
-)
+# set timezone
+timedatectl set-timezone "${BS_SERVER_TIMEZONE}"
 
-ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_SETTINGS_SMTP_SITES}" $BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS \
+DOCUMENT_ROOT="${BS_PATH_DEFAULT_SITE}"
+
+# setup repos
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_SETUP_REPOS}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS"
+
+# install deps
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_INSTALL_DEPS}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS"
+
+# setup bashrc
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_SETUP_BASHRC}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS" \
+  -e "user_server_sites=${BS_USER_SERVER_SITES} \
+  group_user_server_sites=${BS_GROUP_USER_SERVER_SITES} \
+  path_sites=${BS_PATH_SITES} "
+
+# setup postfix
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_SETUP_POSTFIX}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS"
+
+# setup user / nginx / mysql / apache2 / firewalld / php-fpm
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_INITIAL_SETUP}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS" \
+  -e "domain=default \
+
+  db_name=${DB_NAME} \
+  db_user=${DB_USER} \
+  db_password=${DBPASS} \
+  mysql_flavor=${BS_DB_FLAVOR} \
+  mysql_version=${BS_DB_VERSION} \
+  mysql_character_set_server=${BS_DB_CHARACTER_SET_SERVER} \
+  mysql_collation_server=${BS_DB_COLLATION} \
+
+  site_user_password=${site_user_password} \
+
+  path_sites=${BS_PATH_SITES} \
+  document_root=${DOCUMENT_ROOT} \
+
+  delete_files=$(IFS=,; echo "${DELETE_FILES[*]}") \
+
+  download_bitrix_install_files_new_site=$(IFS=,; echo "${BS_DOWNLOAD_BITRIX_INSTALL_FILES_NEW_SITE[*]}") \
+  timeout_download_bitrix_install_files_new_site=${BS_TIMEOUT_DOWNLOAD_BITRIX_INSTALL_FILES_NEW_SITE} \
+
+  user_server_sites=${BS_USER_SERVER_SITES} \
+  group_user_server_sites=${BS_GROUP_USER_SERVER_SITES} \
+
+  permissions_sites_dirs=${BS_PERMISSIONS_SITES_DIRS} \
+  permissions_sites_files=${BS_PERMISSIONS_SITES_FILES} \
+
+  service_nginx_name=${BS_SERVICE_NGINX_NAME} \
+  path_nginx=${BS_PATH_NGINX} \
+  path_nginx_sites_conf=${BS_PATH_NGINX_SITES_CONF} \
+  path_nginx_sites_enabled=${BS_PATH_NGINX_SITES_ENABLED} \
+
+  service_apache_name=${BS_SERVICE_APACHE_NAME} \
+  path_apache=${BS_PATH_APACHE} \
+  path_apache_sites_conf=${BS_PATH_APACHE_SITES_CONF} \
+  path_apache_sites_enabled=${BS_PATH_APACHE_SITES_ENABLED} \
+
+  smtp_path_wrapp_script_sh=${BS_SMTP_PATH_WRAPP_SCRIPT_SH} \
+
+  bx_cron_agents_path_file_after_document_root=${BS_BX_CRON_AGENTS_PATH_FILE_AFTER_DOCUMENT_ROOT} \
+  bx_cron_logs_path_dir=${BS_BX_CRON_LOGS_PATH_DIR} \
+  bx_cron_logs_path_file=${BS_BX_CRON_LOGS_PATH_FILE} \
+
+  push_server_config=${BS_PUSH_SERVER_CONFIG} \
+  php_version=${BX_PHP_DEFAULT_VERSION} \
+  server_timezone=${BS_SERVER_TIMEZONE}"
+
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_SETTINGS_SMTP_SITES}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS" \
   -e "is_new_install_env=Y \
   account_name='' \
   smtp_file_sites_config=${BS_SMTP_FILE_SITES_CONFIG} \
@@ -219,7 +162,7 @@ ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_P
   smtp_file_group_user_log=${BS_SMTP_FILE_GROUP_USER_LOG} \
   smtp_path_wrapp_script_sh=${BS_SMTP_PATH_WRAPP_SCRIPT_SH}"
 
-ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_INSTALL_NEW_FULL_ENVIRONMENT}" $BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS \
+ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_PB_INSTALL_NEW_FULL_ENVIRONMENT}" "$BS_ANSIBLE_RUN_PLAYBOOKS_PARAMS" \
   -e "domain=default \
 
   db_name=${DB_NAME} \
@@ -260,7 +203,31 @@ ansible-playbook "$DEST_DIR_MENU/$DIR_NAME_MENU/ansible/playbooks/${BS_ANSIBLE_P
   bx_cron_logs_path_dir=${BS_BX_CRON_LOGS_PATH_DIR} \
   bx_cron_logs_path_file=${BS_BX_CRON_LOGS_PATH_FILE} \
 
-  push_server_config=${BS_PUSH_SERVER_CONFIG}"
+  push_server_config=${BS_PUSH_SERVER_CONFIG} \
+  
+  php_version=${BX_PHP_DEFAULT_VERSION} \
+  server_timezone=${BS_SERVER_TIMEZONE}"
+
+# disable httpd access logs
+find /etc/apache2/ -type f -print0 | xargs -0 sed -i 's/CustomLog/#CustomLog/g'
+systemctl restart apache2.service
+# fix services
+systemctl restart php"${BX_PHP_DEFAULT_VERSION}"-fpm.service
+systemctl restart postfix@-.service
+firewall-cmd --reload
+apt install -y mysqltuner
+
+# fix journald sizes
+sed -i -e 's/#SystemMaxUse=/SystemMaxUse=100M/g' /etc/systemd/journald.conf
+sed -i -e 's/#RuntimeMaxUse=/RuntimeMaxUse=100M/g' /etc/systemd/journald.conf
+systemctl restart systemd-journald
+
+if [ "$BS_PUSH_SERVER_STOPPED" == true  ]; then
+  systemctl stop push-server.service
+  systemctl disable push-server.service
+  systemctl stop redis.service
+  systemctl disable redis.service
+fi
 
 echo -e "\n\n";
 echo "Full environment installed";
@@ -268,10 +235,3 @@ echo -e "\n";
 echo "Password for the user ${BS_USER_SERVER_SITES}:";
 echo "${site_user_password}";
 echo -e "\n";
-
-END
-
-bash /root/temp_install_full_environment.sh
-
-rm /root/temp_install_full_environment.sh
-rm /root/run.sh
